KOKKOS: remove silent fp32/fp64 conversions in the newly ported styles
Building the KOKKOS package in the mixed and single precision modes with
clang and -Wimplicit-float-conversion -Wdouble-promotion showed that the
styles ported on this branch silently mix KK_FLOAT and double, so parts of
their kernels fall back to fp64 arithmetic on the device.

Make every conversion explicit, following the convention the rest of the
package already uses:

- Bare libm calls on KK_FLOAT arguments resolve to the double overload and
  promote, so they are qualified with Kokkos:: to pick the float overload.
  Where an exponent was an int literal it is made KK_FLOAT, since the integral
  overload of pow() promotes as well.

- Double literals and constants mixed into KK_FLOAT expressions are cast, so
  the arithmetic stays in KK_FLOAT.  These fold at compile time.

- Base class double scalars used in a kernel get one local KK_FLOAT copy with
  the established _kk suffix.

- Accumulators keep KK_ACC_FLOAT and everything else KK_FLOAT, so the cast at
  an accumulator boundary follows the declared view type: widening on the way
  into an EV_FLOAT field or a t_kkacc_ view, narrowing where such a value is
  read back into KK_FLOAT math.  Host side reductions into base class doubles
  such as eng_vdwl and virial[] are cast to double.

- Exchange and communication buffers are double, so the KK_FLOAT values packed
  into and unpacked out of them are converted explicitly in both directions.

Where a style already did its bias or group math in double, as in
compute_temp_ramp, compute_temp_profile and fix_addtorque_group, that precision
is kept and only the boundaries are spelled out, so the arithmetic is unchanged.

Headers are shared by several translation units, and the warnings a header
produces depend on which templates each unit instantiates, so every including
unit was checked in both the single and the mixed precision build.

A full build of all 398 translation units is now free of
-Wimplicit-float-conversion and -Wdouble-promotion warnings in src/KOKKOS in
both the single and the mixed precision build.  In the default double build
every added cast is a no-op.

+++ src/KOKKOS/compute_temp_ramp_kokkos.h
@@ -93,7 +93,7 @@ class ComputeTempRampKokkos : public ComputeTempRamp {
 // NOLINTNEXTLINE
   KOKKOS_INLINE_FUNCTION
   double ramp_bias(const int &i) const {
-    double fraction = (x(i,coord_dim) - coord_lo) / (coord_hi - coord_lo);
+    double fraction = (static_cast<double>(x(i,coord_dim)) - coord_lo) / (coord_hi - coord_lo);
     fraction = (fraction < 0.0) ? 0.0 : fraction;
     fraction = (fraction > 1.0) ? 1.0 : fraction;
     return v_lo + fraction*(v_hi - v_lo);
